Refactor: Modularize FileSelector component
I've decomposed the main FileSelector.tsx component into several smaller, more focused sub-components to improve maintainability and readability.

Key changes:
- Created DirectoryTreeView.tsx: Handles the rendering of the directory structure, file/folder icons, and expansion/selection within the tree.
- Created FileSelectorHeader.tsx: Manages the search input, global action buttons (Select All, Clear), and the smart exclusion suggestion UI.
- Created SelectionSummary.tsx: Displays statistics about the current selection (counts, total size, file types) and lists selected items.
- Created FileSelectorStatusBar.tsx: Shows overall item counts, search result counts, and information about excluded items.
- Moved type definitions (FileSystemItem, FileSelectorProps) to src/types/fileSelector.ts.
- Refactored FileSelector.tsx to be an orchestrator component, managing state and passing data/callbacks to the new sub-components.
- Removed hardcoded sampleFileSystem data, relying on the fileSystemData prop.
- Cleaned up imports and performed minor code hygiene.

This refactoring enhances the component's structure, making it easier to understand, test, and maintain in the future.
