--- src/components/FileSelector.new.tsx
@@ -1,14 +1,0 @@
-// Copy the entire content of FileSelector.tsx here and change the export line to:
-// export function FileSelector({
-//   initialPath = "/",
-//   onSelectionChange,
-//   smartExcludeSuggestions = DEFAULT_SMART_EXCLUDE_SUGGESTIONS,
-//   className,
-//   fileSystemData = sampleFileSystem,
-// }: FileSelectorProps) {
-// ... rest of the component ...
-
-// Then at the end of the file, remove the default export and just have:
-// }
-
-// The file should end with the component's closing brace, no additional exports needed
